+++ src/geometry/line.c
@@ -17,7 +17,7 @@
 #include <polygon_collider/geometry/line.h>
 #include <polygon_collider/geometry/vector.h>
 
-Line line_get(float slope, float intercept)
+Line line_get(double slope, double intercept)
 {
     Line line;
     line.slope = slope;
@@ -36,21 +36,21 @@ Line line_get_from_points(Vector left, Vector right)
 double line_get_segment_angular_mass(Line line, double left, double right)
 {
     double parameter[5];
-    parameter[0] = 0.0f;
-    parameter[1] = line.intercept * line.intercept * line.intercept / 3.0f;
+    parameter[0] = 0.0;
+    parameter[1] = line.intercept * line.intercept * line.intercept / 3.0;
     parameter[2] = line.slope * line.intercept * line.intercept;
     parameter[3] = line.slope * line.slope * line.intercept + line.intercept;
-    parameter[4] = line.slope * line.slope * line.slope / 3.0f + line.slope;
+    parameter[4] = line.slope * line.slope * line.slope / 3.0 + line.slope;
 
-    double left_power = 1.0f;
-    double right_power = 1.0f;
-    double angular_mass = 0.0f;
+    double left_power = 1.0;
+    double right_power = 1.0;
+    double angular_mass = 0.0;
     for(unsigned int i = 1; i <= 4; i++)
     {
         left_power *= left;
         right_power *= right;
 
-        double delta = parameter[i] * (right_power - left_power) / (double)i;
+        double delta = parameter[i] * (right_power - left_power) / i;
         angular_mass += delta;
     }
